+++ src/frontend-pwa/src/components/utility/Greeting/greeting.style.ts
@@ -1,0 +1,16 @@
+import styled from '@emotion/styled';
+
+const GreetingWrapper = styled.p`
+  display: flex;
+  width: 100%;
+  padding: 5pt 10pt;
+  text-align: center;
+  margin: -2em 0 0 0;
+  font-size: 14pt;
+  min-height: 2em;
+  color: #036;
+  justify-content: center;
+  font-style: italic;
+`;
+
+export default GreetingWrapper;
