+++ src/frontend-pwa/src/views/Home/home.styles.ts
@@ -24,19 +24,6 @@ export const Wrapper = styled.div`
     justify-content: center;
 `;
 
-export const GreetingWrapper = styled.p`
-    display: flex;
-    width: 100%;
-    padding: 5pt 10pt;
-    text-align: center;
-    margin: -2em 0 0 0;
-    font-size: 14pt;
-    min-height: 2em;
-    color: #036;
-    justify-content: center;
-    font-style: italic;
-`;
-
 export const ButtonWrapper = styled.div`
     padding: 5px;
 `;
